requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
opencv-python

